+++ docling/datamodel/base_models.py
@@ -117,6 +117,7 @@ class InputFormat(str, Enum):
     LATEX = "latex"
     EMAIL = "email"
     EPUB = "epub"
+    BOXNOTE = "boxnote"
 
 
 class OutputFormat(str, Enum):
@@ -129,6 +130,8 @@ class OutputFormat(str, Enum):
     DOCTAGS = "doctags"
     VTT = "vtt"
     DOCLANG = "doclang"
+    DCLX = "dclx"
+    CHUNKS = "chunks"
 
 
 FormatToExtensions: dict[InputFormat, list[str]] = {
@@ -155,6 +158,7 @@ FormatToExtensions: dict[InputFormat, list[str]] = {
     InputFormat.LATEX: ["tex", "latex"],
     InputFormat.EMAIL: ["eml"],
     InputFormat.EPUB: ["epub"],
+    InputFormat.BOXNOTE: ["boxnote"],
 }
 
 FormatToMimeType: dict[InputFormat, list[str]] = {
@@ -221,6 +225,7 @@ FormatToMimeType: dict[InputFormat, list[str]] = {
     InputFormat.LATEX: ["text/x-tex", "application/x-tex", "text/x-latex"],
     InputFormat.EMAIL: ["message/rfc822"],
     InputFormat.EPUB: ["application/epub+zip"],
+    InputFormat.BOXNOTE: ["application/vnd.box.boxnote"],
 }
 
 MimeTypeToFormat: dict[str, list[InputFormat]] = {